requirements.txt
click==8.1.7
fastapi==0.101.1
openai==0.28.0
pymysql==1.1.0
python-dotenv==1.0.0
starlette==0.27.0
tqdm==4.66.1
urllib3==2.0.4
uvicorn==0.23.2
SQLAlchemy==2.0.20
googlemaps==4.10.0
pika==1.3.2
